Update iP8PlCase_Mex.Rmd
Updates to Mexico code for:
Scraping number of installments.
Several broken seller links on MercadoLibre's end - there are many of them and it is challenging to find a way to identify the HTTP 500 errors and handle them appropriately.  Will require more time.  (I tested Uruguay to see if MercadoLibre is cutting off my IP access but it is not.  Must be a Mexico-specific problem....)