Added scraping and cleaning for color(s) and size(s) of helmet.
Colors and sizes are either displayed in list form or a single option form.  Each form has different nodes.  The single option form is wrapped in new line and tab characters.  If the items are in list form, they are collapsed into a single string where the elements of the list are comma-separated.