Two bug fixes
(1) arrival_time was being removed but some cleaning was being done on it instead of on shipping, which had taken on its values.  arrival_time was also still being displayed on the legend despite being removed.
(2) More seriously, the legend was misrepresenting the values that represent months and years in the actual cleaned data because of a mix up where I had commented out some code and didn't explain why, so I later uncommented it thinking I had meant to put it back in.  Won't happen again.