Exploring alternative to scraping
Not being paid any more but still curious about how I could do this better.

Seems to work in getting search results, but the order of the search results doesn't appear to match the order the appear in on the website (I need to retry when searching from an incognito tab to see if cookies are throwing it off).  Results also don't contain any product rating information (star rating) - would need to find way to supplement that.